scripts/snapshot/requirements.txt
python-bitcoinrpc==1.0
pytz==2024.1
